requirements.txt
SpeechRecognition
pyttsx3
pyaudio
playsound
python-dotenv
groq
Requests
gTTS
pydub